% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictedROC_defaults.R
\docType{data}
\name{default_part_colors}
\alias{default_part_colors}
\title{This code here defines global variables for the package restrictedROC}
\format{
An object of class \code{character} of length 2.
}
\usage{
default_part_colors
}
\description{
This code here defines global variables for the package restrictedROC
}
\keyword{datasets}
