% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_density_values.R
\name{generate_density_values}
\alias{generate_density_values}
\title{Generate values for given densities}
\usage{
generate_density_values(
  ...,
  xmin = -1,
  xmax = 1,
  length.out = 100,
  do_melt = TRUE
)
}
\arguments{
\item{...}{List or multiple arguments of density functions in the following form:
positive = function(x) dnorm(x, mean = 0, sd = 1),
negative = function(x) dnorm(x, mean = -1, sd = 1),
negative2 = function(x) dnorm(x, mean = -3, sd = 1)}

\item{xmin}{Minimum value where the given function(s) are calculated}

\item{xmax}{Maximum value where the given function(s) are calculated}

\item{length.out}{How many values should be generated between xmin and xmax? Used in \code{\link[=seq]{seq()}}}

\item{do_melt}{If TRUE, apply \code{\link[=melt_gendata]{melt_gendata()}}}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{list() or tibble::tibble()
}\if{html}{\out{</div>}}
}
\description{
Generate values for a given list of density functions in the range from \code{xmin} to \code{xmax}
with \code{length.out} elements. Usually not used directly but by \code{\link[=plot_density_ROC]{plot_density_ROC()}}
}
\examples{
generate_density_values(
    positive = function(x) dnorm(x, mean = 0, sd = 1),
    negative = function(x) dnorm(x, mean = -1, sd = 1),
    negative2 = function(x) dnorm(x, mean = -3, sd = 1)
)
generate_density_values(
    positive = function(x) dnorm(x, mean = 0, sd = 1),
    xmin = 0, xmax = 1
)
generate_density_values(
    positive = function(x) dnorm(x, mean = 0, sd = 1),
    do_melt = FALSE
)
generate_density_values(
    positive = function(x) dnorm(x, mean = 0, sd = 1),
    do_melt = TRUE
)
}
