% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_gendata.R
\name{melt_gendata}
\alias{melt_gendata}
\title{Melt generated data}
\usage{
melt_gendata(df)
}
\arguments{
\item{df}{List or data.frame of values.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{data.frame with two columns:
	"Distribution"	Contains the name/colname of the respective value
	"Value" 		The value itself
}\if{html}{\out{</div>}}
}
\description{
Melt data generated in the form of lists. Given a list or data.frame of elements,
melt it from
columnA columnB columnC
into
Distribution Value
columnA			xA_1
columnA			xA_2
...				...
columnB			xB_2
...				...
columnC			xC_2
...				...
}
\examples{
a <- list("A" = 1:3, "B" = 5:19)
melt_gendata(a)

a <- data.frame("a" = 1:10, "b" = 6:15)
melt_gendata(a)
}
