% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_applied.R
\name{merge_applied}
\alias{merge_applied}
\title{Merge (cbind) applied rROC results}
\usage{
merge_applied(
  rroc_applied,
  which_preds = c("all", "all_split", "predictor", "bounded", "keep")
)
}
\arguments{
\item{rroc_applied}{A \emph{single outcome} result of apply_restriction()}

\item{which_preds}{Which predictions should be returned? Default is all.
You can also specify "all_split" to return a list of predictions for each
type of prediction (e.g. "predictor", "bounded", "keep"). Or you can specify
a subset of these types (e.g. c("predictor", "bounded")), then the bound predictions
including the predictions of only these types will be returned.}
}
\description{
Merge (cbind) applied rROC results
}
