% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_ROC.R
\name{perf_pROC}
\alias{perf_pROC}
\title{Performance ROC curves (pROC)}
\usage{
perf_pROC(
  response,
  predictor,
  quiet = FALSE,
  coords_ret = c("tp", "fp", "tpr", "fpr", "threshold"),
  ...
)
}
\arguments{
\item{response}{A vector containing the true class labels. Care that it is VERY important
which class is the positive class because the \emph{predictions} are ordered according
to \code{restriction}}

\item{predictor}{A vector containing the predictions.}

\item{quiet}{Parameter of \link[pROC:roc]{pROC::roc}}

\item{coords_ret}{Coordinates from \code{\link[pROC:coords]{pROC::coords()}} which should be returned in the "perf"-listelement}

\item{...}{Further parameters to \link[pROC:roc]{pROC::roc}}
}
\value{
List of
perf:			Result of \code{\link[=perf_pROC]{perf_pROC()}}, comes from \code{\link[pROC:coords]{pROC::coords()}}
auc:			Area under the ROC curve, comes from \code{\link[pROC:auc]{pROC::auc()}}
positive_label: Label of the positive class
}
\description{
Performance ROC curves (pROC)
}
\examples{

data(aSAH, package = "pROC")
restrictedROC:::perf_pROC(
    response = aSAH$outcome,
    predictor = aSAH$ndka
)
restrictedROC:::perf_pROC(
    response = aSAH$outcome,
    predictor = aSAH$ndka,
    levels = c("Poor", "Good"), # response-values for NEGATIVE and POSITIVE
    direction = "<" # predictions of positives are higher than negatives
)
restrictedROC:::perf_pROC(
    response = aSAH$outcome,
    predictor = aSAH$ndka,
    levels = c("Poor", "Good"), # response-values for NEGATIVE and POSITIVE
    direction = ">" # predictions of positives are higher than negatives
)
restrictedROC:::perf_pROC(
    response = aSAH$outcome,
    predictor = aSAH$ndka,
    levels = c("Good", "Poor"), # response-values for NEGATIVE and POSITIVE
    direction = "<" # predictions of positives are higher than negatives
)

}
