% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_ROC.R
\name{plot_density_ROC}
\alias{plot_density_ROC}
\title{Plot density + ROC}
\usage{
plot_density_ROC(
  density_positive,
  cdf_positive,
  density_negative,
  quantile_negative,
  length.out = 500,
  xmin = -3,
  xmax = 3,
  colors_pos_neg_both = colors_pos_neg_both_default
)
}
\arguments{
\item{density_positive}{\code{function(x) {}}: Returns the density given a specific value x.
Refers to values coming from the "positive"-class.}

\item{cdf_positive}{\code{function(q) {...}}: Returns the cumulative distribution value given a specific quantile \code{q}.
Refers to values coming from the "positive"-class.}

\item{density_negative}{\code{function(x) {}}: Returns the density given a specific value x.
Refers to values coming from the "negative"-class.}

\item{quantile_negative}{\code{function(x) {}}: Returns the density given a specific value x.
Refers to values coming from the "negative"-class.}

\item{length.out}{The number of simulated points for the plots from the distributions.
\code{length.out} points will be generated between \verb{[xmin, xmax]}.
PURELY VISUAL, ROC curves are calculated based on the actual distributions!}

\item{xmin}{Minimum value of simulated points for the plots from the distributions.
\code{length.out} points will be generated between \verb{[xmin, xmax]}.
PURELY VISUAL, ROC curves are calculated based on the actual distributions!}

\item{xmax}{Maximum value of simulated points for the plots from the distributions.
\code{length.out} points will be generated between \verb{[xmin, xmax]}.
PURELY VISUAL, ROC curves are calculated based on the actual distributions!}

\item{colors_pos_neg_both}{Colors for positive ("+"), negative ("-") and overlapping "+/-" areas of the
densities}
}
\value{
\code{patchwork}'ed ggplots: left density, right ROC curve.
}
\description{
Given positive and negative distributions, plot the corresponding densities and their ROC-curves.
}
\examples{
plot_density_ROC(
    density_positive = function(x) dnorm(x, mean = 0, sd = 1),
    cdf_positive = function(x) pnorm(x, mean = 0, sd = 1),
    density_negative = function(x) dnorm(x, mean = -1, sd = 1.5),
    quantile_negative = function(x) qnorm(x, mean = -1, sd = 1.5),
    xmin = -4, xmax = 7
)
}
