% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_ROC_empirical.R
\name{plot_density_ROC_empirical}
\alias{plot_density_ROC_empirical}
\title{Plot density + ROC}
\usage{
plot_density_ROC_empirical(
  values_grouped,
  length.out = 500,
  xmin = -3,
  xmax = 3,
  ...
)
}
\arguments{
\item{values_grouped}{List-like element. (also data.frames or matrices) where the elements are the
samples from the different distributions.
E.g.:
# A tibble: 1,000 × 2
dist_1 dist_2
\if{html}{\out{<dbl>}}  \if{html}{\out{<dbl>}}
1  0.974  0.257
2  0.196 -0.780
3 -0.125 -0.264
4  0.701  0.260}

\item{length.out}{See \code{\link[=plot_density_empirical]{plot_density_empirical()}}. Granularity of density plot (pure visualization).}

\item{xmin}{See \code{\link[=plot_density_empirical]{plot_density_empirical()}}.}

\item{xmax}{See \code{\link[=plot_density_empirical]{plot_density_empirical()}}.}

\item{...}{Further arguments to \code{\link[=perf_ROC]{perf_ROC()}}}
}
\value{
\code{patchwork}'ed ggplots: left density, right ROC curve.
}
\description{
Given positive and negative distributions, plot the corresponding densities and their ROC-curves.
}
\examples{

sim_samples <- sim(
    list(
        "dist_1" = function(x) rnorm(x, mean = 1, sd = 1),
        "dist_2" = function(x) rnorm(x, mean = 0, sd = 1)
    ),
    do_melt = FALSE,
    length.out = 1000
)
plot_density_ROC_empirical(sim_samples)
}
