% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_ROC.R
\name{plot_density_ROC_str}
\alias{plot_density_ROC_str}
\title{Plot density + ROC based on stats::_\if{html}{\out{<NAME>}}}
\usage{
plot_density_ROC_str(
  dist_positive_str = "norm",
  dist_negative = "norm",
  dist_positive_args = list(mean = 0, sd = 1),
  dist_negative_args = list(mean = 0, sd = 1),
  length.out = 500,
  xmin = -3,
  xmax = 3
)
}
\arguments{
\item{dist_positive_str}{A name of any distribution of the \verb{stats::} package. E.g. "norm", "binom", "exp".
This is the distribution of values coming from the "positive"-class.
For all possible distributions check ?stats::Distributions}

\item{dist_negative}{A name of any distribution of the \verb{stats::} package. E.g. "norm", "binom", "exp".
This is the distribution of values coming from the "negative"-class.
For all possible distributions check ?stats::Distributions}

\item{dist_positive_args}{Arguments to \code{dist_positive_str} as named list.}

\item{dist_negative_args}{Arguments to \code{dist_negative_str} as named list.}

\item{length.out}{The number of simulated points for the plots from the distributions.
\code{length.out} points will be generated between \verb{[xmin, xmax]}.
PURELY VISUAL, ROC curves are calculated based on the actual distributions!}

\item{xmin}{Minimum value of simulated points for the plots from the distributions.
\code{length.out} points will be generated between \verb{[xmin, xmax]}.
PURELY VISUAL, ROC curves are calculated based on the actual distributions!}

\item{xmax}{Maximum value of simulated points for the plots from the distributions.
\code{length.out} points will be generated between \verb{[xmin, xmax]}.
PURELY VISUAL, ROC curves are calculated based on the actual distributions!}
}
\value{
\code{patchwork}'ed ggplots: left density, right ROC curve.
}
\description{
Given positive and negative distribution \emph{NAMES} and their arguments,
plot the corresponding densities and their ROC-curves.
}
\examples{

plot_density_ROC_str()
plot_density_ROC_str(
    dist_positive_str = "norm",
    dist_negative = "norm",
    dist_positive_args = list("mean" = 0, "sd" = 1),
    dist_negative_args = list("mean" = 1, "sd" = 1),
    xmin = -4, xmax = 7
)
# pdf("removeme.pdf", width = 14, height = 6)
plot_density_ROC_str(
    dist_positive_str = "norm",
    dist_negative = "norm",
    dist_positive_args = list("mean" = 1, "sd" = 1),
    dist_negative_args = list("mean" = 0, "sd" = 1),
    xmin = -4, xmax = 7
)
# dev.off()
# pdf("removeme_2.pdf", width = 14, height = 6)
plot_density_ROC_str(
    dist_positive_str = "norm",
    dist_negative = "norm",
    dist_positive_args = list("mean" = 1, "sd" = .5),
    dist_negative_args = list("mean" = 1, "sd" = 1),
    xmin = -4, xmax = 7, length.out = 2503
)
# dev.off()
# pdf("removeme_3.pdf", width = 14, height = 6)
plot_density_ROC_str(
    dist_positive_str = "norm",
    dist_negative = "norm",
    dist_positive_args = list("mean" = 1, "sd" = 2),
    dist_negative_args = list("mean" = 1, "sd" = 1),
    xmin = -4, xmax = 7, length.out = 2503
)
# dev.off()
tmp <- plot_density_ROC_str(
    dist_positive_str = "norm",
    dist_negative = "norm",
    dist_positive_args = list("mean" = 1, "sd" = 2),
    dist_negative_args = list("mean" = 1, "sd" = 1),
    xmin = -4, xmax = 7, length.out = 2503
)

}
