% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_ROC_empirical.R
\name{plot_density_rROC_empirical}
\alias{plot_density_rROC_empirical}
\title{Plot 4 rROC plots}
\usage{
plot_density_rROC_empirical(
  values_grouped,
  length.out = 500,
  xmin = NA,
  xmax = NA,
  single_rROC = NULL,
  positive_label = NULL,
  direction = "<",
  part_colors = default_part_colors,
  plot_thresholds = TRUE,
  plot_thresholds_fpr = TRUE,
  plot_n_points = NA
)
}
\arguments{
\item{values_grouped}{List-like element. (also data.frames or matrices) where the elements are the
samples from the different distributions.
E.g.:
# A tibble: 1,000 × 2
dist_1 dist_2
\if{html}{\out{<dbl>}}  \if{html}{\out{<dbl>}}
1  0.974  0.257
2  0.196 -0.780
3 -0.125 -0.264
4  0.701  0.260}

\item{length.out}{See \code{\link[=plot_density_empirical]{plot_density_empirical()}}. Granularity of density plot (pure visualization).}

\item{xmin}{See \code{\link[=plot_density_empirical]{plot_density_empirical()}}.}

\item{xmax}{See \code{\link[=plot_density_empirical]{plot_density_empirical()}}.}

\item{single_rROC}{Optional, if not given, rROC will be calculated on the fly.}

\item{positive_label}{Positive label for calculating ROC curve and the densities. If NULL, will be
automatically generated from the data.}

\item{direction}{See \code{\link[pROC:roc]{pROC::roc()}}, but only "<" implemented.}

\item{part_colors}{Default: default_part_colors
A vector c("high"=COLOR_A, "low"=COLOR_B) with colors for high-value (markerHIGH)
and low-value (markerLOW) parts of calculating the restricted ROC.}

\item{plot_thresholds}{Plot the optimal restrictions for high/low part and the max_total (best overall)}

\item{plot_thresholds_fpr}{Plot the max_total threshold on plots B, C and D}
}
\value{
\code{patchwork}'ed ggplots, see Description for their meaning.
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A, Top-left:		Given positive and negative distributions, plot the corresponding densities.
B, Top-right:		Corresponding ROC curve
C, Bottom-left:		False positive rate VS threshold
D, Bottom-right:	rzAUC plot for markerHIGH and markerLOW restricted parts, the rzAUC is shown
}\if{html}{\out{</div>}}
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{For further description see our publication on restricted ROC curves.
}\if{html}{\out{</div>}}
}
\examples{

sim_samples <- sim(
    list(
        "dist_1" = function(x) rnorm(x, mean = 1, sd = 1),
        "dist_2" = function(x) rnorm(x, mean = 0, sd = 1)
    ),
    do_melt = FALSE,
    length.out = 50
)
plot_density_rROC_empirical(sim_samples)
plot_density_rROC_empirical(sim_samples, positive_label = "dist_1")
plot_density_rROC_empirical(sim_samples, positive_label = "dist_2")
plot_density_rROC_empirical(sim_samples, part_colors = c("high" = "green", "low" = "yellow"))
plot_density_rROC_empirical(sim_samples, plot_thresholds = FALSE)
plot_density_rROC_empirical(sim_samples, plot_thresholds_fpr = FALSE)



tmp <- simple_rROC(
    response = melt_gendata(sim_samples)[["Distribution"]],
    predictor = melt_gendata(sim_samples)[["Value"]],
    direction = "<",
    positive_label = "dist_1",
    return_proc = TRUE
)
single_rROC <- simple_rROC_interpret(tmp)
plot_density_rROC_empirical(sim_samples, single_rROC = single_rROC)

}
