% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distributions.R
\name{plot_distributions_2}
\alias{plot_distributions_2}
\title{Plot exactly 2 distributions}
\usage{
plot_distributions_2(
  df,
  colors_pos_neg_both = c(`+` = "#008B45FF", `+/-` = "#3B4992FF", `-` = "#EE0000FF"),
  name_dist_1 = "Positive",
  name_dist_2 = "Negative"
)
}
\arguments{
\item{df}{Table with the distribution values.
E.g.:
# A tibble: 1,000 × 2
Positive  Negative
\if{html}{\out{<dbl>}}     \if{html}{\out{<dbl>}}
1  0.974     0.257
2  0.196    -0.780
3 -0.125    -0.264
4  0.701     0.260}

\item{colors_pos_neg_both}{Colors of positive (+), negative (-) and overlapping (+/-) density-areas}

\item{name_dist_1}{Name of the first distribution (first column)}

\item{name_dist_2}{Name of the second distribution (second column)}
}
\value{
ggplot object
}
\description{
Plot exactly 2 distributions
}
\examples{
# See e.g. plot_density_ROC_empirical()
}
