% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ROC_theoretical.R
\name{plot_ROC_theoretical}
\alias{plot_ROC_theoretical}
\title{ROC curve for positive cumulative distribution and negative quantile function}
\usage{
plot_ROC_theoretical(pnorm_positive, qnorm_negative, length.out = 500)
}
\arguments{
\item{pnorm_positive}{Cumulative distribution function of positive class}

\item{qnorm_negative}{Quantile function (inverse of the cumulative distribution function) of the negative class}

\item{length.out}{How many false positive rates should the ROC-curve be calculated for?}
}
\value{
List of
- ggplot of the ROC curve
- AUC of the curve
}
\description{
ROC curve for positive cumulative distribution and negative quantile function
}
\examples{

plot_ROC_theoretical(
    pnorm_positive = function(x) pnorm(x, mean = 1, sd = 1),
    qnorm_negative = function(x) qnorm(x, mean = 1, sd = .5)
)
}
