% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rROC.R
\name{plot_rROC}
\alias{plot_rROC}
\title{plot ROC including AUC significance}
\usage{
plot_rROC(
  x,
  col_rzAUC = "#999999",
  part = c("high", "low"),
  part_colors = default_part_colors,
  split_roc_score = FALSE
)
}
\arguments{
\item{x}{rROC result from \code{\link{simple_rROC}}, \code{\link{simple_rROC_interpret}} or \code{\link{rROC}}.}

\item{col_rzAUC}{Color for rzAUC points if split_roc_score=FALSE}

\item{part}{"high", "low" or multiple}

\item{part_colors}{Colors for the \code{part}}

\item{split_roc_score}{Shoult the ROC score (rzAUC) be a separate plot ("TRUE") or inside the ROC
curve plot ("FALSE")}
}
\value{
List of multiple ggplot elements, always "roc",
if split_roc_score=TRUE also a single "rzAUC" plot
}
\description{
Plot the complete ROC curve and the corresponding AUCs.
}
\examples{
# See also test-plot_rROC()
data(aSAH, package = "pROC")
simple_rROC_res <- simple_rROC(
    response = aSAH$outcome,
    predictor = aSAH$ndka
)
single_rROC <- simple_rROC_interpret(simple_rROC_res)
plot_rROC(
    single_rROC,
    split_roc_score = TRUE
)
plot_rROC(
    single_rROC,
    split_roc_score = FALSE
)
plot_rROC(
    single_rROC,
    split_roc_score = FALSE,
    part = "high"
)
plot_rROC(
    single_rROC,
    split_roc_score = FALSE,
    part = "low"
)

}
