% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rROC_part.R
\name{plot_rROC_part}
\alias{plot_rROC_part}
\title{Plot (restricted) ROC curves}
\usage{
plot_rROC_part(
  x,
  p_full_density_ROC = NA,
  threshold = NA,
  fpr = NA,
  color_high = default_part_colors["high"],
  color_low = default_part_colors["low"],
  include_part_auc_text = FALSE
)
}
\arguments{
\item{x}{rROC result including full_pROC.}

\item{p_full_density_ROC}{Density plot from plot_density_ROC_empirical of the full data, if not given it is calculated from rROC_res}

\item{threshold}{At which threshold should be split into high and low. Alternatively, give a false positive rate (fpr)}

\item{fpr}{A false positive rate at which (approximately) the cutoff will be made}

\item{color_high}{Color for high part}

\item{color_low}{Color for low part}

\item{include_part_auc_text}{Should the text-annotations be added or not?}
}
\value{
\link{patchwork}ed plots, see the description.
}
\description{
Plot for every tpr/fpr combination:
\itemize{
\item Full density
\item Full ROC
+ selection rectangles (orange + blue)
+ Full AUC
+ orange+blue partial AUCs
+ scaling factor
+ name rectangle sides
+ orange+blue partial AUCs SCALED
\item ROCs for selections
+ Full AUCs
+ Color rectangle around the plot
\item Densities for selections
}
}
\examples{

library(restrictedROC)
data(aSAH, package = "pROC")
ret_procs <- simple_rROC(
    response = aSAH$outcome,
    predictor = aSAH$ndka,
    return_proc = TRUE
)
# pdf("removeme.pdf")
print(plot_rROC_part(ret_procs, fpr = .5))
# dev.off()
}
