% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rROC.R
\name{plot_rzAUCs}
\alias{plot_rzAUCs}
\title{Plot rzAUC vs FPR}
\usage{
plot_rzAUCs(
  rROC_result,
  part = c("high", "low"),
  part_colors = default_part_colors
)
}
\arguments{
\item{rROC_result}{rROC result from:

\if{html}{\out{<div class="sourceCode">}}\preformatted{    single_rROC <- simple_rROC_interpret(simple_rROC(
    response = aSAH$outcome,
    predictor = aSAH$ndka
    ))
}\if{html}{\out{</div>}}}

\item{part}{vector which parts should be shown, can be multiple}

\item{part_colors}{Named vector which part should receive which color}
}
\value{
ggplot object with rzAUCs shown for all false positive rates. For one FPR
multiple values can exist.
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Plot rzAUCs for all calculated parts
}\if{html}{\out{</div>}}
}
\examples{
# See plot_rROC()
}
