% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.restrictedROC.R
\name{predict.restrictedROC}
\alias{predict.restrictedROC}
\title{S3 class predict for class "restrictedROC"}
\usage{
\method{predict}{restrictedROC}(
  object,
  newdata,
  newdata_predictor_column = 1,
  newdata_response_column = 2,
  pred_high_label = 1,
  pred_low_label = 0,
  original.response = NA,
  original.predictor = NA,
  ...
)
}
\arguments{
\item{object}{Result from \code{\link[=simple_rROC_interpret]{simple_rROC_interpret()}}, easiest calculated from simple_rROC(..., return_proc=TRUE)}

\item{newdata}{The new dataframe which should be predicted}

\item{newdata_predictor_column}{Column name or number of the predictor (What you use to predict the response)}

\item{newdata_response_column}{Column name or number of the true response (/outcome)}

\item{pred_high_label}{Label of predictions HIGHER than the cutoff (Youden-index)}

\item{pred_low_label}{Label of predictions LOWER OR EQUAL than the cutoff (Youden-index)}

\item{original.response}{If the original response cannot be loaded from \code{object}, you have to give the
original response.}

\item{original.predictor}{If the original predictor cannot be loaded from \code{object}, you have to give the
predictor}

\item{...}{Not used but keeps devtools::check() happy}
}
\value{
List of
table_full
Table of full predictions, classified using the Youden-Index. This is the
classical way of predicting.
table_restricted
Table of restricted predictions, classified using the Youden-Index.
Samples which were identified as "unclassifiable" with our restriction
method were removed.
pred
tibble with
"response": 	True response
"predictor":	Values to use as predictors
"keep": 		Should they be kept according to restriction
"prediction":	Youden-based predictions, regardless if kept or not
threshold_and_restriction
Vector with
"restriction":
The restriction value, decides together with \code{restriction_part}
which samples are kept.
"classification_threshold":
Values above will be classified as \code{pred_high_label}
restriction_part
Should "high", "low" or "global" (=all) samples be kept, therefore
samples with values higher or lower (or all) with respect to the restriction
are kept.

\if{html}{\out{<div class="sourceCode">}}\preformatted{tab_classifiable
	Restricted table including the unclassifiable predictions, e.g.

	        	        response
		prediction      	Poor Good
			Poor              29   35
			Good               8   34
			unclassifiable     4    3
}\if{html}{\out{</div>}}
}
\description{
Predict new samples for a given restrictedROC result
}
\examples{

data(aSAH, package = "pROC")
tmp <- simple_rROC(
    response = aSAH$outcome,
    predictor = aSAH$ndka,
    return_proc = TRUE
)
single_rROC <- simple_rROC_interpret(tmp)

predict(
    single_rROC,
    newdata = aSAH,
    newdata_predictor_column = "ndka",
    newdata_response_column = "outcome",
    pred_high_label = "Poor",
    pred_low_label = "Good"
)

tmp <- simple_rROC(
    response = aSAH$outcome,
    predictor = aSAH$ndka
)
single_rROC_noFullROC <- simple_rROC_interpret(tmp)
\dontrun{
predict(
    single_rROC_noFullROC,
    newdata = aSAH,
    newdata_predictor_column = "ndka",
    newdata_response_column = "outcome",
    pred_high_label = "Poor",
    pred_low_label = "Good"
)
}
predict(
    single_rROC_noFullROC,
    newdata = aSAH,
    newdata_predictor_column = "ndka",
    newdata_response_column = "outcome",
    pred_high_label = "Poor",
    pred_low_label = "Good",
    original.response = aSAH$outcome,
    original.predictor = aSAH$ndka
)

}
