% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rROC_model.R
\name{predict_rROC_h2o}
\alias{predict_rROC_h2o}
\title{Predict new data with a trained model}
\usage{
predict_rROC_h2o(
  h2o_model,
  x_prepared,
  y,
  init_h2o = FALSE,
  calculate_metrics = TRUE,
  sample_split = NULL
)
}
\arguments{
\item{h2o_model}{The model to use for prediction.}

\item{x_prepared}{A data frame containing the preprocessed data to use in the
model. The data frame can contain more variables than necessary within h2o_model}

\item{y}{The true outcome variable. Usually a factor.}

\item{init_h2o}{Whether to initialize h2o. Defaults to FALSE as I expect it was
initialized during training already.}

\item{calculate_metrics}{Whether to calculate metrics. Defaults to TRUE.}

\item{sample_split}{A vector of length(y) containing the "split" of the data.
The metrices are calculated for each split separately. Defaults to NULL, which means
that the whole data is used.}
}
\description{
A convenience function to predict new data with a trained model.
The function is a wrapper around the h2o.predict function.
It returns a list with the predictions and optionally the metrics.
}
