% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_modeldata.R
\name{prepare_modeldata.data.frame}
\alias{prepare_modeldata.data.frame}
\title{Prepare model data}
\usage{
\method{prepare_modeldata}{data.frame}(
  x,
  y = NULL,
  rroc_result = NULL,
  rroc_savefile = NULL,
  which_preds = "bounded",
  removed_impute = -1,
  ...
)
}
\arguments{
\item{x}{Data which should be prepared. Can be a data.frame or matrix, but
must have column names. Rows are samples, columns are features.}

\item{y}{The outcome vector. If NULL, rroc_result must be supplied. If given, only
rROC() uses it.}

\item{rroc_result}{The result of rROC(). If NULL, rROC() is called with x, y and \code{...}.}

\item{rroc_savefile}{If not NULL, the rroc_result is saved to this file.}

\item{which_preds}{Which predictions should be returned? Default is "bounded". See
\code{\link{merge_applied}} for details.}

\item{removed_impute}{The value to impute for samples outside the informative range.}

\item{...}{Arguments passed to rROC() if rroc_result is NULL.}
}
\description{
This function prepares the data for the model. It takes the rROC result and
applies the restrictions to the data. It then merges the predictions into a
single data.frame which can be used as input for model building.
}
