% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_modeldata.R
\name{prepare_modeldata}
\alias{prepare_modeldata}
\title{Prepare model data}
\usage{
prepare_modeldata(x, ...)
}
\arguments{
\item{x}{Data which should be prepared. Can be a data.frame or matrix, but
must have column names. Rows are samples, columns are features.}

\item{...}{
  Arguments passed on to \code{\link[=prepare_modeldata.data.frame]{prepare_modeldata.data.frame}}
  \describe{
    \item{\code{y}}{The outcome vector. If NULL, rroc_result must be supplied. If given, only
rROC() uses it.}
    \item{\code{rroc_result}}{The result of rROC(). If NULL, rROC() is called with x, y and \code{...}.}
    \item{\code{rroc_savefile}}{If not NULL, the rroc_result is saved to this file.}
    \item{\code{which_preds}}{Which predictions should be returned? Default is "bounded". See
\code{\link{merge_applied}} for details.}
  }}
}
\description{
This function prepares the data for the model. It takes the rROC result and
applies the restrictions to the data. It then merges the predictions into a
single data.frame which can be used as input for model building.
}
