% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictedROC-package.R
\docType{package}
\name{restrictedROC-package}
\alias{restrictedROC}
\alias{restrictedROC-package}
\title{restrictedROC: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ggrlab.github.io/restrictedROC/}
}

}
\author{
\strong{Maintainer}: Gunther Glehr \email{gunthergl@gmx.net} (\href{https://orcid.org/0000-0002-1495-9162}{ORCID})

}
\keyword{internal}
