% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rROC.R
\name{rROC.data.frame}
\alias{rROC.data.frame}
\title{Restriction for multiple dependent and independent variables}
\usage{
\method{rROC}{data.frame}(
  x,
  dependent_vars,
  independent_vars,
  save_path = NULL,
  parallel_permutations = TRUE,
  n_permutations = 10000,
  save_intermediate = TRUE,
  load_existing_intermediate = TRUE,
  positive_label = 1,
  verbose = TRUE,
  do_plots = FALSE,
  fix_seed = TRUE,
  ...
)
}
\arguments{
\item{x}{A data.frame containing all dependent and independent variables as columns.}

\item{dependent_vars}{A character vector of dependent variable column names.}

\item{independent_vars}{A character vector of independent variable column names.
If NULL, all columns except dependent_vars are used.}

\item{save_path}{Path to save the results to. Intermediate results are saved into
the directory file.path(save_path, "_partial_directory").}

\item{parallel_permutations}{boolean: If TRUE, the permutation will be done via \code{future.apply::future_lapply},
otherwise by \code{base::lapply}}

\item{n_permutations}{How many permutations should be done}

\item{save_intermediate}{Should intermediate results be saved to disk? If TRUE,
every combination by itself is saved into file.path(save_path, "_partial_directory").}

\item{load_existing_intermediate}{Should the earlier saved intermediate results
in the folder file.path(save_path, "_partial_directory") be loaded?}

\item{positive_label}{Label for the positive class. All other values of \code{response} are regarded as negative cases.}

\item{verbose}{Should progress be printed?}

\item{do_plots}{Should the plot_density_rROC_empirical be calculated and returned?}

\item{fix_seed}{boolean: If TRUE, the seed for each permutation will be set by set.seed(permutation_i)}

\item{...}{
  Arguments passed on to \code{\link[=simple_rROC_permutation]{simple_rROC_permutation}}
  \describe{
    \item{\code{return_proc}}{\enumerate{
\item Should pROC::roc() be returned for the full dataset?
2) Should pROC::roc() be returned on each of the part datasets? Only works with
\code{get_all_aucs_fun=get_all_aucs} after  \code{get_all_aucs_norecalculation()} does
not calculate the ROC curves for each restriction separately.
}}
  }}
}
\description{
Restriction for multiple dependent and independent variables.
Traverses all dependent variables and within all independent variables.
Then calculates the rROC in the sense of dependent ~ independent.

Can save intermediate results to disk, to avoid re-calculating for crash-reasons
or to save time when re-running the same analysis.

Can return plot_density_rROC_empirical for every combination.
}
