% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rROC.R
\name{rROC.numeric}
\alias{rROC.numeric}
\title{rROC on a single numeric vector}
\usage{
\method{rROC}{numeric}(x, y, ...)
}
\arguments{
\item{x}{Numeric vector, values of the independent variable for every element/sample.}

\item{y}{Factor vector, values of the dependent variable for every element/sample.}

\item{...}{
  Arguments passed on to \code{\link[=rROC.data.frame]{rROC.data.frame}}
  \describe{
    \item{\code{save_path}}{Path to save the results to. Intermediate results are saved into
the directory file.path(save_path, "_partial_directory").}
    \item{\code{save_intermediate}}{Should intermediate results be saved to disk? If TRUE,
every combination by itself is saved into file.path(save_path, "_partial_directory").}
    \item{\code{load_existing_intermediate}}{Should the earlier saved intermediate results
in the folder file.path(save_path, "_partial_directory") be loaded?}
    \item{\code{do_plots}}{Should the plot_density_rROC_empirical be calculated and returned?}
    \item{\code{verbose}}{Should progress be printed?}
    \item{\code{n_permutations}}{How many permutations should be done}
    \item{\code{fix_seed}}{boolean: If not FALSE, the seed for each permutation will be set by set.seed(fix_seed + permutation_i)}
    \item{\code{parallel_permutations}}{boolean: If TRUE, the permutation will be done via \code{future.apply::future_lapply},
otherwise by \code{base::lapply}}
    \item{\code{positive_label}}{Label for the positive class. All other values of \code{response} are regarded as negative cases.}
  }}
}
\description{
rROC on a single numeric vector
}
