% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rROC.R
\name{rROC}
\alias{rROC}
\title{Restriction for multiple dependent and independent variables}
\usage{
rROC(x, ...)
}
\arguments{
\item{x}{\itemize{
\item {data.frame}: {
See \code{\link{rROC.data.frame}}. data.frame containing all dependent and
independent variables as columns. Dependent/independent variable column names
must be given as "dependent_vars"/"independent_vars" arguments.
}
\item {matrix}: {
See \code{\link{rROC.matrix}}. Matrix of (samples x features).
Dependent variable(s) must be given as "y" argument.
}
\item {numeric vector}: {
See \code{\link{rROC.numeric}}. Numeric vector of independent variable.
Dependent variable(s) must be given as "y" argument.
}
}}

\item{...}{
  Arguments passed on to \code{\link[=rROC.data.frame]{rROC.data.frame}}, \code{\link[=simple_rROC_permutation]{simple_rROC_permutation}}
  \describe{
    \item{\code{save_path}}{Path to save the results to. Intermediate results are saved into
the directory file.path(save_path, "_partial_directory").}
    \item{\code{save_intermediate}}{Should intermediate results be saved to disk? If TRUE,
every combination by itself is saved into file.path(save_path, "_partial_directory").}
    \item{\code{load_existing_intermediate}}{Should the earlier saved intermediate results
in the folder file.path(save_path, "_partial_directory") be loaded?}
    \item{\code{do_plots}}{Should the plot_density_rROC_empirical be calculated and returned?}
    \item{\code{verbose}}{Should progress be printed?}
    \item{\code{n_permutations}}{How many permutations should be done}
    \item{\code{fix_seed}}{boolean: If not FALSE, the seed for each permutation will be set by set.seed(fix_seed + permutation_i)}
    \item{\code{parallel_permutations}}{boolean: If TRUE, the permutation will be done via \code{future.apply::future_lapply},
otherwise by \code{base::lapply}}
    \item{\code{positive_label}}{Label for the positive class. All other values of \code{response} are regarded as negative cases.}
    \item{\code{return_proc}}{\enumerate{
\item Should pROC::roc() be returned for the full dataset?
2) Should pROC::roc() be returned on each of the part datasets? Only works with
\code{get_all_aucs_fun=get_all_aucs} after  \code{get_all_aucs_norecalculation()} does
not calculate the ROC curves for each restriction separately.
}}
  }}
}
\description{
Restriction for multiple dependent and independent variables
}
