% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Simulate values from distributions}
\usage{
sim(..., length.out = 100, do_melt = TRUE)
}
\arguments{
\item{...}{Every argument must be a function which takes the number of values which should
be generated, e.g.:

\if{html}{\out{<div class="sourceCode">}}\preformatted{    `function(n) random_n_values_from_distribution(n)`
}\if{html}{\out{</div>}}}

\item{length.out}{How many samples should be drawn from each population}

\item{do_melt}{TRUE:
# A tibble: 200 × 2
Distribution  Value
\if{html}{\out{<chr>}}         \if{html}{\out{<dbl>}}
1 negative      0.805
2 positive      2.62
3 negative      1.19
4 positive      1.49

FALSE:
# A tibble: 100 × 2
negative positive
\if{html}{\out{<dbl>}}    \if{html}{\out{<dbl>}}
1   0.631     2.81
2   0.687     1.87
3   0.0347    1.61
4  -1.15      0.529}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{See do_melt parameter
}\if{html}{\out{</div>}}
}
\description{
Simulate values from distributions
}
\examples{

sim(
    negative = function(x) rnorm(x, mean = 0, sd = 1),
    positive = function(x) rnorm(x, mean = 1, sd = 1)
)
sim(
    list(
        negative = function(x) rnorm(x, mean = 0, sd = 1),
        positive = function(x) rnorm(x, mean = 1, sd = 1)
    )
)
sim(
    negative = function(x) rnorm(x, mean = 0, sd = 1),
    positive = function(x) rnorm(x, mean = 1, sd = 1),
    do_melt = FALSE
)
sim(
    negative = function(x) rnorm(x, mean = 0, sd = 1),
    positive = function(x) rnorm(x, mean = 1, sd = 1),
    do_melt = TRUE
)

sim(
    negative = function(x) rnorm(x, mean = 0, sd = 1),
    positive = function(x) rnorm(x, mean = 1, sd = 1),
    length.out = 10
)

}
