% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_rROC_interpret.R
\name{simple_rROC_interpret}
\alias{simple_rROC_interpret}
\title{restricted ROC interpretation}
\usage{
simple_rROC_interpret(perf_rroc, use_abs_rzAUC = TRUE)
}
\arguments{
\item{perf_rroc}{Result from simple_rROC}

\item{use_abs_rzAUC}{One of the results is the restriction with the maximum rzAUC value.
If TRUE, report the absolute highest, otherwise the general highest
including the sign.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{S3-class restrictedROC of
 - performances
     previously reported perf_rroc result, previously seen as "joined_aucs"
 - global
     AUC, AUC variance under H0, rzAUC and asymptotic pvalue

     when including all samples. This pvalue is now appropriately usable according to
 - keep_highs
     AUC, AUC variance under H0, rzAUC, asymptotic pvalue, threshold

     Threshold is the restriction value, all other values the corresponding
     performances for the highest (potentially absolute) rzAUC when keeping high
     values (markerHIGH)
 - keep_lows
     AUC, AUC variance under H0, rzAUC, asymptotic pvalue, threshold

     Threshold is the restriction value, all other values the corresponding
     performances for the highest (potentially absolute) rzAUC when keeping low
     values (markerLOW)
 - max_total
     AUC, AUC variance under H0, rzAUC, asymptotic pvalue, threshold, part

     Threshold is the restriction value, all other values the corresponding
     performances for the highest (potentially absolute) rzAUC overall.
     This compares "global", "keep_highs" and "keep_lows", therefore `part`
     denotes where the result was coming from.
 - positive_label
     What is the positive label
 - pROC_full
     [pROC::roc()] result of the full dataset, only if `return_pROC` was TRUE
}\if{html}{\out{</div>}}
}
\description{
Interpret the results of restriction, e.g. make them nicer to read and further
process.
}
\examples{
data(aSAH, package = "pROC")
tmp <- simple_rROC(
    response = aSAH$outcome,
    predictor = aSAH$ndka
)
simple_rROC_interpret(tmp)
simple_rROC_interpret(tmp, use_abs_rzAUC = FALSE)

}
