% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_rROC_permutation.R
\name{simple_rROC_permutation}
\alias{simple_rROC_permutation}
\title{restricted ROC}
\usage{
simple_rROC_permutation(
  response,
  predictor,
  positive_label = NULL,
  direction = "<",
  n_permutations = 100,
  fix_seed = TRUE,
  parallel_permutations = FALSE,
  return_proc = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{response}{A vector containing the true class labels. Care that it is VERY important
which class is the positive class because the \emph{predictions} are ordered according
to \code{restriction}}

\item{predictor}{A vector containing the predictions.}

\item{positive_label}{Label for the positive class. All other values of \code{response} are regarded as negative cases.}

\item{direction}{See \code{\link[pROC:roc]{pROC::roc()}}, but only "<" is implemented right now. Maybe changing the
positive_label already solves your problem.}

\item{n_permutations}{How many permutations should be done}

\item{fix_seed}{boolean: If TRUE, the seed for each permutation will be set by set.seed(permutation_i)}

\item{parallel_permutations}{boolean: If TRUE, the permutation will be done via \code{future.apply::future_lapply},
otherwise by \code{base::lapply}}

\item{return_proc}{\enumerate{
\item Should pROC::roc() be returned for the full dataset?
2) Should pROC::roc() be returned on each of the part datasets? Only works with
\code{get_all_aucs_fun=get_all_aucs} after  \code{get_all_aucs_norecalculation()} does
not calculate the ROC curves for each restriction separately.
}}

\item{verbose}{Print in which permutation we are}
}
\value{
List of:
- Results of \code{\link[=simple_rROC_interpret]{simple_rROC_interpret()}}:
performances, global, keep_highs, keep_lows, max_total, positive_label
"pROC_full": Result of \code{\link[pROC:roc]{pROC::roc()}} on the full dataset, calculated once
"permutation_pval":
Permutation p-values globally and with optimized ("max") restriction

\if{html}{\out{<div class="sourceCode">}}\preformatted{			pval.twoside.max pval.twoside.global      n_permutations
						0.16                0.01              100.00
	"perm_max_bound":
		Bound table of the optimal restriction results for all permutations

         # A tibble: 100 × 6
         auc auc_var_H0 rzAUC pval_asym threshold part
         <dbl>      <dbl> <dbl>     <dbl>     <dbl> <chr>
         	1 0.161     0.0134  -2.92   0.00345     17.4  high
         2 0.631     0.00635  1.64   0.101       11.7  high
         3 0.332     0.00826 -1.85   0.0644      13.2  high

	"perm_global_bound":
		Bound table of the global AUC (= use all samples) for all permutations

         # A tibble: 100 × 4
              auc auc_var_H0   rzAUC pval_asym
            <dbl>      <dbl>   <dbl>     <dbl>
          1 0.453      0.453 -0.827      0.408
          2 0.504      0.504  0.0627     0.950
}\if{html}{\out{</div>}}
}
\description{
Calculate the restricted ROC curves.
}
\examples{

data(aSAH, package = "pROC")
a <- simple_rROC_permutation(
    response = aSAH$outcome,
    predictor = aSAH$ndka
)
}
