% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.rROC}
\alias{summary.rROC}
\title{Give a summary of a rROC() result}
\usage{
\method{summary}{rROC}(
  object,
  relevant_cols_regex = c("pval.twoside.*", "n_permutations", "positive_label",
    ".*\\\\.auc$", "part", "restriction", "informative_range.*"),
  current_level = 0,
  ...
)
}
\arguments{
\item{object}{A \code{\link{rROC}} result}

\item{relevant_cols_regex}{A regex to filter the columns of a single restriction.}

\item{current_level}{The current level of the nested list. This is used internally.}

\item{...}{Additional arguments passed to \code{\link{summary}}.}
}
\value{
A summary of the result:
}
\description{
\code{\link{rROC}} returns a nested list. It looks something like:
\code{single_restriction <- reslist[[dependent_variables]][[independent_variables]]}
where every single restriction has done permutations, a p-value and additional info.
This function gives a summary of the result.
}
