% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rROC_model.R
\name{train_rROC_h2o}
\alias{train_rROC_h2o}
\title{Train a predictive model using (rROC-preprocessed) data}
\usage{
train_rROC_h2o(
  x_prepared,
  y,
  init_h2o = TRUE,
  h2o_trainfun = function(df, col_y, cols_x, ...) {
     h2o.randomForest(training_frame
    = df, y = col_y, x = cols_x, ntrees = 1000, max_depth = 20, min_rows = 1, nbins = 20,
    seed = 4242, ...)
 },
  ...
)
}
\arguments{
\item{x_prepared}{A data frame containing the preprocessed data to use in the
model. The data frame should not contain the outcome variable.
Usually the result of a call to \code{\link{prepare_modeldata}}.}

\item{y}{The outcome variable. Usually a factor.}

\item{init_h2o}{Whether to initialize h2o. Defaults to TRUE.}

\item{h2o_trainfun}{The function to use to train the model. Defaults to
\code{\link{h2o.randomForest}} with some default parameters.
Needs to have \code{df}, \code{col_y} and \code{cols_x} as arguments where \code{df} is
the data frame to use, \code{col_y} is the column index of the outcome variable
and \code{cols_x} is a vector of column indices of the features to use to predict the
outcome.}

\item{...}{Arguments passed to h2o_trainfun.}
}
\description{
A convenience function to train a model using the data x_prepared to
predict the outcome y. Per default, the function is a wrapper around the
h2o.randomForest function.
}
