% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_restriction.R
\name{apply_restriction}
\alias{apply_restriction}
\title{Apply a (nested list) of rROC results to new data}
\usage{
apply_restriction(object, newdata, feature = NA, removed_impute = -1)
}
\arguments{
\item{object}{A (nested list) of rROC results. See \code{\link{rROC}} for details.}

\item{newdata}{A data frame with all predictor variables.}

\item{feature}{The current feature. If \code{object} is a single restrictedROC object,
this is the name of the feature. If \code{object} is a (nested list), this is the name
of the current list element. Example: \code{list("outcome_1"=list("feature_1"=rroc_result))}
\code{feature} is \code{"feature_1"} when \code{rroc_result} is processed when
rroc_result is detected.}

\item{removed_impute}{The value to impute for samples outside the informative range.}
}
\value{
A list with three elements:
\itemize{
\item \code{predictions}: A data frame with the original predictor values and the
imputed values. It has three columns:
\describe{
\item{\code{predictor}}{is the original predictor value.}
\item{\code{keep}}{is a logical vector that indicates whether the sample is inside the
informative range.}
\item{\code{bounded}}{is the imputed value.}
\item{\code{bounded_twoside}}{
is the imputed value for two-sided restriction: RestrictionValue + or - \code{removed_impute}
}
}
\item \code{thresholds}: A data frame with the feature, threshold, the part of the
predictor that was used and the imputed value for predictor values outside the
informative range. The informative range is then:
\describe{
\item{\code{low}}{\code{(-Inf, threshold)}}
\item{\code{high}}{\code{(threshold, Inf)}}
}
}
}
\description{
Restriction returns an informative range which tells that some samples should be
discarded according to their predictor value. This function applies the restriction
to a new dataset. Predictor values outside the informative range are imputed with
\code{removed_impute}.
}
