% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rROC.R
\name{rROC.matrix}
\alias{rROC.matrix}
\title{rROC on a matrix of (samples x features)}
\usage{
\method{rROC}{matrix}(x, y, ...)
}
\arguments{
\item{x}{Matrix, rows are samples, columns are features. Every feature should be restricted.}

\item{y}{Factor vector, values of the dependent variable for every column (element/sample).}

\item{...}{
  Arguments passed on to \code{\link[=rROC.data.frame]{rROC.data.frame}}
  \describe{
    \item{\code{save_path}}{Path to save the results to. Intermediate results are saved into
the directory file.path(save_path, "_partial_directory").}
    \item{\code{save_intermediate}}{Should intermediate results be saved to disk? If TRUE,
every combination by itself is saved into file.path(save_path, "_partial_directory").}
    \item{\code{load_existing_intermediate}}{Should the earlier saved intermediate results
in the folder file.path(save_path, "_partial_directory") be loaded?}
    \item{\code{do_plots}}{Should the plot_density_rROC_empirical be calculated and returned?}
    \item{\code{verbose}}{Should progress be printed?}
    \item{\code{n_permutations}}{How many permutations should be done}
    \item{\code{fix_seed}}{boolean: If not FALSE, the seed for each permutation will be set by set.seed(fix_seed + permutation_i)}
    \item{\code{parallel_permutations}}{boolean: If TRUE, the permutation will be done via \code{future.apply::future_lapply},
otherwise by \code{base::lapply}}
    \item{\code{positive_label}}{Label for the positive class. All other values of \code{response} are regarded as negative cases.}
  }}
}
\value{
A list of lists of simple_rROC_permutation and plot results. It is structured as follows:
\itemize{
\item {dependent variable}:
\itemize{
\item {independent variable}:
\describe{
\item{"plots"}{\code{\link{plot_density_rROC_empirical}} result}
\item{"permutation"}{\code{\link{simple_rROC_permutation}} result}
}
}
}
}
\description{
rROC on a matrix of (samples x features)
}
